useResponsiveness: hook to return the width of the window
