+++ src/components/menu/MenuWidget.tsx
@@ -5,6 +5,7 @@ import ButtonHelp from "./ButtonHelp";
 import SearchField from "./SearchField";
 import RoutingResults from "./RoutingResults";
 import { Feature } from "../types";
+import useResponsiveness from "./useResponsiveness";
 
 const geoLocationOptions: PositionOptions = {
   enableHighAccuracy: true,
@@ -28,23 +29,14 @@ type Props = {
  * The Menu. In Widget mode.
  */
 const MenuWidget = (props: Props) => {
-  const [prevWidth, setPrevWidth] = useState(window.innerWidth);
-  const [searchFieldsOpen, setSearchFieldsOpen] = useState(
-    window.innerWidth >= 450
-  );
   const [renderFormKeys, setRenderFormKeys] = useState(true);
+  const prevWidth = useResponsiveness();
+  const [searchFieldsOpen, setSearchFieldsOpen] = useState(
+    window.innerWidth >= 460
+  );
   const [waitingForGeolocation, setWaitingForGeolocation] = useState(false);
 
-  useEffect(() => {
-    const updateBySize = () => {
-      if (prevWidth !== window.innerWidth) {
-        setPrevWidth(window.innerWidth);
-        setSearchFieldsOpen(window.innerWidth >= 460);
-      }
-    };
-    window.addEventListener("resize", updateBySize);
-    return () => window.removeEventListener("resize", updateBySize);
-  }, [prevWidth]);
+  useEffect(() => setSearchFieldsOpen(prevWidth >= 460), [prevWidth]);
 
   const resetRoute = () => {
     props.reset();
